Don't connect on __construct(). Add connect().
